+++ CHANGELOG.md
@@ -1,3 +1,8 @@
-## 1.0.0
+## 0.1.1
+	- Add embed reply
+	- fix build1.1 renamed to build1 problem
+	- add google drive support (folders will be zipped)
+
+## 0.1.0
 
 - Initial version.